+++ 0_dump/train_ODE.py
@@ -14,7 +14,7 @@ from lib.supervised.sgg.dsgdetr.track import get_sequence_with_tracking
 from lib.supervised.sgg.dsgdetr.matcher import HungarianMatcher
 
 from train_base import fetch_train_basic_config, prepare_optimizer, fetch_loss_functions
-from lib.supervised.sga import ODE as ODE
+from lib.supervised.sga import SceneSayerODE as ODE
 
 
 def train_ode():
